Setup basic routes & pages, complete login page
