+++ package.json
@@ -15,6 +15,8 @@
     "framer-motion": "^11.2.6",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-icons": "^5.2.1",
+    "react-loader-spinner": "^6.1.6",
     "react-router-dom": "^6.23.1"
   },
   "devDependencies": {
