+++ src/App.css
@@ -1,10 +1,14 @@
-html {
-  height: 100%;
-  width: 100%;
-  margin: 0;
-  padding-left: 1rem;
-  padding-right: 1rem;
-  font-family: sans-serif;
+#root {
+  max-width: 1280px;
+  margin: 0 auto;
+}
+
+body {
+  background: #0c2f0a; /* fallback color */
+  background: linear-gradient(135deg, rgb(7, 117, 136), rgb(1, 7, 41));
+
+  min-width: 320px;
+  min-height: 100vh;
 }
 
 .str-video {
@@ -16,4 +20,11 @@ html {
   flex-direction: column;
   min-width: 0;
   max-width: 100%;
+}
+
+.galindo-regular {
+  font-family: "Galindo", sans-serif;
+  font-weight: 400;
+  font-style: normal;
+  font-size: 90px;
 }
