+++ src/index.css
@@ -3,12 +3,12 @@
 @tailwind utilities;
 
 :root {
-  font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
+  font-family: Noto Sans, Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
   font-weight: 400;
 
   color-scheme: light dark;
-  color: rgba(24, 24, 24, 0.87);
+  color: rgba(253, 253, 253, 0.87);
   background-color: #18191c;
 
   font-synthesis: none;
