+++ src/pages/Dashboard.tsx
@@ -1,0 +1,6 @@
+const Dashboard = () => {
+  return (
+    <div>Dashboard</div>
+  )
+}
+export default Dashboard
